requirements.txt
amqp==2.6.1
attrs==21.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==4.4.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coverage==5.5
Django==2.2.24
django-admin-sortable2==1.0
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
djangorestframework==3.12.4
future==0.18.2
importlib-metadata==4.8.0
iniconfig==1.1.1
kombu==4.6.11
packaging==21.0
pluggy==0.13.1
prompt-toolkit==3.0.20
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.2
python-memcached==1.59
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
vine==1.3.0
wcwidth==0.2.5
zipp==3.5.0
